pyproject.toml
[tool.poetry]
name = "algotutor"
version = "0.1.0"
description = "Algorithm Tutor Platform"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "app"}]


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
motor = "^3.3.1"
beanie = "^1.23.1"
pydantic = "^2.4.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
docker = "^6.1.3"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
google-auth = "^2.23.4"
google-auth-oauthlib = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.1"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
